Move console printing listeners to reporting package
Originally the console printing listeners were in the console package
and were package private. Making them public and moving them to a more
generic package makes them useful to a wider audience.

Implements #1963.
