+++ junit-platform-launcher/src/main/java/org/junit/platform/launcher/TestIdentifier.java
@@ -130,7 +130,7 @@ public final class TestIdentifier implements Serializable {
 	 *
 	 * @return the legacy reporting name; never {@code null} or blank
 	 * @see org.junit.platform.engine.TestDescriptor#getLegacyReportingName()
-	 * @see org.junit.platform.launcher.listeners.LegacyReportingUtils
+	 * @see org.junit.platform.reporting.legacy.LegacyReportingUtils
 	 */
 	public String getLegacyReportingName() {
 		return this.legacyReportingName;
