+++ platform-tests/src/test/java/org/junit/platform/reporting/console/VerboseTreeListenerTests.java
@@ -8,16 +8,14 @@
  * https://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junit.platform.console.tasks;
+package org.junit.platform.reporting.console;
 
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.platform.console.tasks.FlatPrintingListener.INDENTATION;
+import static org.junit.jupiter.api.Assertions.assertLinesMatch;
 import static org.junit.platform.engine.TestExecutionResult.failed;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
 import org.junit.platform.engine.UniqueId;
@@ -26,9 +24,9 @@ import org.junit.platform.fakes.TestDescriptorStub;
 import org.junit.platform.launcher.TestIdentifier;
 
 /**
- * @since 1.0
+ * @since 1.3.2
  */
-class FlatPrintingListenerTests {
+class VerboseTreeListenerTests {
 
 	private static final String EOL = System.lineSeparator();
 
@@ -38,11 +36,14 @@ class FlatPrintingListenerTests {
 		listener(stringWriter).executionSkipped(newTestIdentifier(), "Test" + EOL + "disabled");
 		String[] lines = lines(stringWriter);
 
-		assertEquals(3, lines.length);
-		assertAll("lines in the output", //
-			() -> assertEquals("Skipped:     demo-test ([engine:demo-engine])", lines[0]), //
-			() -> assertEquals(INDENTATION + "=> Reason: Test", lines[1]), //
-			() -> assertEquals(INDENTATION + "disabled", lines[2]));
+		assertLinesMatch(List.of( //
+			"+-- %c ool test", //
+			"     tags: []", //
+			" uniqueId: [engine:demo-engine]", //
+			"   parent: []", //
+			"   reason: Test", //
+			"             disabled", //
+			"   status: [S] SKIPPED"), List.of(lines));
 	}
 
 	@Test
@@ -51,11 +52,7 @@ class FlatPrintingListenerTests {
 		listener(stringWriter).reportingEntryPublished(newTestIdentifier(), ReportEntry.from("foo", "bar"));
 		String[] lines = lines(stringWriter);
 
-		assertEquals(2, lines.length);
-		assertAll("lines in the output", //
-			() -> assertEquals("Reported:    demo-test ([engine:demo-engine])", lines[0]), //
-			() -> assertTrue(lines[1].startsWith(INDENTATION + "=> Reported values: ReportEntry [timestamp =")), //
-			() -> assertTrue(lines[1].endsWith(", foo = 'bar']")));
+		assertLinesMatch(List.of("  reports: ReportEntry \\[timestamp = .+, foo = 'bar'\\]"), List.of(lines));
 	}
 
 	@Test
@@ -64,17 +61,30 @@ class FlatPrintingListenerTests {
 		listener(stringWriter).executionFinished(newTestIdentifier(), failed(new AssertionError("Boom!")));
 		String[] lines = lines(stringWriter);
 
-		assertAll("lines in the output", //
-			() -> assertEquals("Finished:    demo-test ([engine:demo-engine])", lines[0]), //
-			() -> assertEquals(INDENTATION + "=> Exception: java.lang.AssertionError: Boom!", lines[1]));
+		assertLinesMatch(List.of("   caught: java.lang.AssertionError: Boom!", //
+			">> STACKTRACE >>", //
+			" duration: \\d+ ms", //
+			"   status: [X] FAILED"), List.of(lines));
 	}
 
-	private FlatPrintingListener listener(StringWriter stringWriter) {
-		return new FlatPrintingListener(new PrintWriter(stringWriter), true);
+	@Test
+	void failureMessageWithFormatSpecifier() {
+		StringWriter stringWriter = new StringWriter();
+		listener(stringWriter).executionFinished(newTestIdentifier(), failed(new AssertionError("%crash")));
+		String[] lines = lines(stringWriter);
+
+		assertLinesMatch(List.of("   caught: java.lang.AssertionError: %crash", //
+			">> STACKTRACE >>", //
+			" duration: \\d+ ms", //
+			"   status: [X] FAILED"), List.of(lines));
+	}
+
+	private VerboseTreePrintingListener listener(StringWriter stringWriter) {
+		return new VerboseTreePrintingListener(new PrintWriter(stringWriter), true, 16, Theme.ASCII);
 	}
 
 	private static TestIdentifier newTestIdentifier() {
-		TestDescriptorStub testDescriptor = new TestDescriptorStub(UniqueId.forEngine("demo-engine"), "demo-test");
+		TestDescriptorStub testDescriptor = new TestDescriptorStub(UniqueId.forEngine("demo-engine"), "%c ool test");
 		return TestIdentifier.from(testDescriptor);
 	}
